+++ addons/account/controllers/__init__.py
@@ -3,3 +3,4 @@
 
 from . import portal
 from . import terms
+from . import download_edi_docs
