+++ addons/account_edi/__init__.py
@@ -1,5 +1,4 @@
 # -*- encoding: utf-8 -*-
 
 from . import models
-from . import controller
 from . import wizard
